+++ flax/linen/summary.py
@@ -13,23 +13,30 @@
 # limitations under the License.
 
 """Flax Module summary library."""
-from abc import ABC, abstractmethod
+
+from abc import ABC
+from abc import abstractmethod
 import dataclasses
 import io
 from types import MappingProxyType
 from typing import Any, Callable, Dict, Iterable, List, Mapping, Optional, Sequence, Set, Tuple, Type, Union
-from flax.core import unfreeze
 
-import flax.linen.module as module_lib
 from flax.core import meta
-from flax.core.scope import CollectionFilter, DenyList, FrozenVariableDict, MutableVariableDict
+from flax.core import unfreeze
+from flax.core.scope import CollectionFilter
+from flax.core.scope import DenyList
+from flax.core.scope import FrozenVariableDict
+from flax.core.scope import LazyRng
+from flax.core.scope import MutableVariableDict
+import flax.linen.module as module_lib
 import jax
 import jax.numpy as jnp
+import numpy as np
 import rich.console
 import rich.table
 import rich.text
 import yaml
-import numpy as np
+
 
 PRNGKey = Any  # pylint: disable=invalid-name
 RNGSequences = Dict[str, PRNGKey]
@@ -93,6 +100,9 @@ class Row:
 
   Attributes:
     path: A tuple of strings that represents the path to the module.
+    module_type: type of the Module.
+    method: method of the module called.
+    inputs: inputs to the module.
     outputs: Output of the Module as reported by `capture_intermediates`.
     module_variables: Dictionary of variables in the module (no submodules
       included).
@@ -101,6 +111,8 @@ class Row:
       then this field is the same as `module_variables`, however if a
       summarization is done then this dictionary potentially contains parameters
       from submodules depending on the depth of the Module in question.
+    flops: FLOPs cost of calling the module method.
+    vjp_flops: FLOPs cost of calling the VJP of the module method.
   """
 
   path: Tuple[str, ...]
@@ -110,6 +122,8 @@ class Row:
   outputs: Any
   module_variables: Dict[str, Dict[str, Any]]
   counted_variables: Dict[str, Dict[str, Any]]
+  flops: int
+  vjp_flops: int
 
   def __post_init__(self):
     self.inputs = self.inputs
@@ -160,19 +174,22 @@ def tabulate(
     console_kwargs: Optional[Mapping[str, Any]] = None,
     table_kwargs: Mapping[str, Any] = MappingProxyType({}),
     column_kwargs: Mapping[str, Any] = MappingProxyType({}),
+    compute_flops: bool = False,
+    compute_vjp_flops: bool = False,
     **kwargs,
 ) -> Callable[..., str]:
   """Returns a function that creates a summary of the Module represented as a table.
 
-  This function accepts most of the same arguments and internally calls `Module.init`,
-  except that it returns a function of the form `(*args, **kwargs) -> str` where `*args`
-  and `**kwargs` are passed to `method` (e.g. `__call__`) during the forward pass.
+  This function accepts most of the same arguments and internally calls
+  `Module.init`, except that it returns a function of the form
+  `(*args, **kwargs) -> str` where `*args` and `**kwargs` are passed to
+  `method` (e.g. `__call__`) during the forward pass.
 
-  `tabulate` uses `jax.eval_shape` under the hood to run the forward computation without
-  consuming any FLOPs or allocating memory.
+  `tabulate` uses `jax.eval_shape` under the hood to run the forward computation
+  without consuming any FLOPs or allocating memory.
 
-  Additional arguments can be passed into the `console_kwargs` argument, for example,
-  `{'width': 120}`. For a full list of `console_kwargs` arguments, see:
+  Additional arguments can be passed into the `console_kwargs` argument, for
+  example, `{'width': 120}`. For a full list of `console_kwargs` arguments, see:
   https://rich.readthedocs.io/en/stable/reference/console.html#rich.console.Console
 
   Example::
@@ -188,38 +205,45 @@ def tabulate(
         return nn.Dense(2)(h)
 
     x = jnp.ones((16, 9))
-    tabulate_fn = nn.tabulate(Foo(), jax.random.key(0))
+    tabulate_fn = nn.tabulate(
+        Foo(), jax.random.key(0), compute_flops=True, compute_vjp_flops=True)
 
     print(tabulate_fn(x))
 
 
   This gives the following output::
 
-                                    Foo Summary
-    ┏━━━━━━━━━┳━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━━━━━━┓
-    ┃ path    ┃ module ┃ inputs        ┃ outputs       ┃ params               ┃
-    ┡━━━━━━━━━╇━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━━━━━━┩
-    │         │ Foo    │ float32[16,9] │ float32[16,2] │                      │
-    ├─────────┼────────┼───────────────┼───────────────┼──────────────────────┤
-    │ Dense_0 │ Dense  │ float32[16,9] │ float32[16,4] │ bias: float32[4]     │
-    │         │        │               │               │ kernel: float32[9,4] │
-    │         │        │               │               │                      │
-    │         │        │               │               │ 40 (160 B)           │
-    ├─────────┼────────┼───────────────┼───────────────┼──────────────────────┤
-    │ Dense_1 │ Dense  │ float32[16,4] │ float32[16,2] │ bias: float32[2]     │
-    │         │        │               │               │ kernel: float32[4,2] │
-    │         │        │               │               │                      │
-    │         │        │               │               │ 10 (40 B)            │
-    ├─────────┼────────┼───────────────┼───────────────┼──────────────────────┤
-    │         │        │               │         Total │ 50 (200 B)           │
-    └─────────┴────────┴───────────────┴───────────────┴──────────────────────┘
+                                           Foo Summary
+    ┏━━━━━━━━━┳━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━━━━━━━━━┳━━━━━━━┳━━━━━━━━━━━┳━━━━━━━━━━━━━━━━━┓
+    ┃ path    ┃ module ┃ inputs        ┃ outputs       ┃ flops ┃ vjp_flops ┃ params          ┃
+    ┡━━━━━━━━━╇━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━━━━━━━━━╇━━━━━━━╇━━━━━━━━━━━╇━━━━━━━━━━━━━━━━━┩
+    │         │ Foo    │ float32[16,9] │ float32[16,2] │ 1504  │ 4460      │                 │
+    ├─────────┼────────┼───────────────┼───────────────┼───────┼───────────┼─────────────────┤
+    │ Dense_0 │ Dense  │ float32[16,9] │ float32[16,4] │ 1216  │ 3620      │ bias:           │
+    │         │        │               │               │       │           │ float32[4]      │
+    │         │        │               │               │       │           │ kernel:         │
+    │         │        │               │               │       │           │ float32[9,4]    │
+    │         │        │               │               │       │           │                 │
+    │         │        │               │               │       │           │ 40 (160 B)      │
+    ├─────────┼────────┼───────────────┼───────────────┼───────┼───────────┼─────────────────┤
+    │ Dense_1 │ Dense  │ float32[16,4] │ float32[16,2] │ 288   │ 840       │ bias:           │
+    │         │        │               │               │       │           │ float32[2]      │
+    │         │        │               │               │       │           │ kernel:         │
+    │         │        │               │               │       │           │ float32[4,2]    │
+    │         │        │               │               │       │           │                 │
+    │         │        │               │               │       │           │ 10 (40 B)       │
+    ├─────────┼────────┼───────────────┼───────────────┼───────┼───────────┼─────────────────┤
+    │         │        │               │               │       │     Total │ 50 (200 B)      │
+    └─────────┴────────┴───────────────┴───────────────┴───────┴───────────┴─────────────────┘
 
-                          Total Parameters: 50 (200 B)
+                                   Total Parameters: 50 (200 B)
 
 
   **Note**: rows order in the table does not represent execution order,
   instead it aligns with the order of keys in `variables` which are sorted
   alphabetically.
+
+  **Note**: `vjp_flops` returns `0` if the module is not differentiable.
 
   Args:
     module: The module to tabulate.
@@ -229,21 +253,31 @@ def tabulate(
       depth limit, its parameter count and bytes will be added to the row of its
       first shown ancestor such that the sum of all rows always adds up to the
       total number of parameters of the Module.
+    show_repeated: If `True`, repeated calls to the same module will be shown
+      in the table, otherwise only the first call will be shown. Default is
+      `False`.
     mutable: Can be bool, str, or list. Specifies which collections should be
       treated as mutable: ``bool``: all/no collections are mutable. ``str``: The
       name of a single mutable collection. ``list``: A list of names of mutable
       collections. By default all collections except 'intermediates' are
       mutable.
-    show_repeated: If `True`, repeated calls to the same module will be shown
-      in the table, otherwise only the first call will be shown. Default is
-      `False`.
-    console_kwargs: An optional dictionary with additional keyword arguments that
-      are passed to `rich.console.Console` when rendering the table. Default arguments
-      are `{'force_terminal': True, 'force_jupyter': False}`.
+    console_kwargs: An optional dictionary with additional keyword arguments
+      that are passed to `rich.console.Console` when rendering the table.
+      Default arguments are `{'force_terminal': True, 'force_jupyter': False}`.
     table_kwargs: An optional dictionary with additional keyword arguments that
       are passed to `rich.table.Table` constructor.
     column_kwargs: An optional dictionary with additional keyword arguments that
-      are passed to `rich.table.Table.add_column` when adding columns to the table.
+      are passed to `rich.table.Table.add_column` when adding columns to the
+      table.
+    compute_flops: whether to include a `flops` column in the table listing the
+      estimated FLOPs cost of each module forward pass. Does incur actual
+      on-device computation / compilation / memory allocation, but still
+      introduces overhead for large modules (e.g. extra 20 seconds for a
+      Stable Diffusion's UNet, whereas otherwise tabulation would finish in 5
+      seconds).
+    compute_vjp_flops: whether to include a `vjp_flops` column in the table
+      listing the estimated FLOPs cost of each module backward pass. Introduces
+      a compute overhead of about 2-3X of `compute_flops`.
     **kwargs: Additional arguments passed to `Module.init`.
 
   Returns:
@@ -258,21 +292,139 @@ def tabulate(
 
   def _tabulate_fn(*fn_args, **fn_kwargs):
     table_fn = _get_module_table(
-        module, depth=depth, show_repeated=show_repeated
+        module,
+        depth=depth,
+        show_repeated=show_repeated,
+        compute_flops=compute_flops,
+        compute_vjp_flops=compute_vjp_flops,
     )
+
     table = table_fn(rngs, *fn_args, **fn_kwargs, **kwargs)
-    return _render_table(table, console_kwargs, table_kwargs, column_kwargs)
+
+    non_param_cols = [
+        'path',
+        'module',
+        'inputs',
+        'outputs',
+    ]
+
+    if compute_flops:
+      non_param_cols.append('flops')
+    if compute_vjp_flops:
+      non_param_cols.append('vjp_flops')
+
+    return _render_table(
+        table, console_kwargs, table_kwargs, column_kwargs, non_param_cols
+    )
 
   return _tabulate_fn
+
+
+def _get_flops(fn, *args, **kwargs):
+  e = jax.jit(fn).lower(*args, **kwargs)
+  cost = e.cost_analysis()
+  flops = int(cost['flops']) if 'flops' in cost else 0
+  return flops
+
+
+def _get_call_flops(
+    c: module_lib._CallInfo,
+    compute_flops: bool,
+    compute_vjp_flops: bool,
+) -> tuple[int, int]:
+  """Return the FLOPs of executing the call `c` in the call stack.
+
+  Does not perform actual computation / compilation / memory allocation, but
+  still introduces overhead for large modules.
+
+  Args:
+    c: ``_CallInfo``.
+    compute_flops: whether to compute forward pass FLOPs. Return `-1` otherwise.
+    compute_vjp_flops: whether to compute backward pass FLOPs. Return `-1`
+      otherwise.
+
+  Returns:
+    FLOPs of executing forward pass of `c`, and its VJP.
+  """
+
+  if not compute_flops and not compute_vjp_flops:
+    return -1, -1
+
+  rngs = jax.tree_map(
+      lambda x: x.rng, c.rngs, is_leaf=lambda x: isinstance(x, LazyRng)
+  )
+
+  args = jax.tree_map(_from_value_representation, c.args)
+  kwargs = jax.tree_map(_from_value_representation, c.kwargs)
+
+  leaves, treedef = jax.tree_util.tree_flatten((args, kwargs))
+  dynamic_leaves = []
+  dynamic_idxs = []
+  for i, arg in enumerate(leaves):
+    if isinstance(arg, jax.ShapeDtypeStruct):
+      dynamic_leaves.append(arg)
+      dynamic_idxs.append(i)
+
+  def _get_inputs(dynamic_leaves):
+    new_leaves: list[Any] = leaves.copy()
+    for i, arg in zip(dynamic_idxs, dynamic_leaves):
+      new_leaves[i] = arg
+    return treedef.unflatten(new_leaves)
+
+  def init(rngs, dynamic_leaves):
+    """`c.module.init` closed over static keyword arguments."""
+    args, kwargs = _get_inputs(dynamic_leaves)
+    return c.module.init(
+        rngs,
+        *args,
+        method=c.method,
+        mutable=c.mutable,
+        **kwargs,
+    )
+
+  variables = jax.eval_shape(init, rngs, dynamic_leaves)
+
+  def apply(variables, rngs, dynamic_leaves):
+    """`c.module.apply` closed over static keyword arguments."""
+    args, kwargs = _get_inputs(dynamic_leaves)
+    return c.module.apply(
+        variables,
+        *args,
+        rngs=rngs,
+        method=c.method,
+        mutable=c.mutable,
+        **kwargs,
+    )
+
+  # Forward pass FLOPs
+  if compute_flops:
+    flops = _get_flops(apply, variables, rngs, dynamic_leaves)
+  else:
+    flops = -1
+
+  if compute_vjp_flops:
+    # Backward pass FLOPs
+    def apply_vjp(variables, rngs, dynamic_leaves):
+      """VJP of `c.module.apply` closed over static keyword arguments."""
+      out, vjp_fn = jax.vjp(apply, variables, rngs, dynamic_leaves)
+      return vjp_fn(out)
+
+    vjp_flops = _get_flops(apply_vjp, variables, rngs, dynamic_leaves)
+  else:
+    vjp_flops = -1
+
+  return flops, vjp_flops
 
 
 def _get_module_table(
     module: module_lib.Module,
     depth: Optional[int],
     show_repeated: bool,
+    compute_flops: bool,
+    compute_vjp_flops: bool,
 ) -> Callable[..., Table]:
-  """A function that takes a Module and returns function with the same signature as `init`
-  but returns the Table representation of the Module."""
+  """A function that takes a Module and returns function with the same signature
+  as `init` but returns the Table representation of the Module."""
 
   def _get_table_fn(*args, **kwargs):
     with module_lib._tabulate_context():
@@ -314,12 +466,13 @@ def _get_module_table(
       rows.append(
           Row(
               c.path,
-              c.module_type,
+              type(c.module),
               c.method,
               inputs,
               c.outputs,
               module_vars,
               counted_vars,
+              *_get_call_flops(c, compute_flops, compute_vjp_flops),
           )
       )
 
@@ -361,8 +514,9 @@ def _get_module_variables(
 def _get_path_variables(
     path: Tuple[str, ...], variables: FrozenVariableDict
 ) -> MutableVariableDict:
-  """A function that takes a path and a variables structure and returns the variable structure at
-  that path."""
+  """A function that takes a path and a variables structure and returns the
+  variable structure at that path.
+  """
   path_variables = {}
 
   for collection in variables:
@@ -380,8 +534,9 @@ def _get_path_variables(
 
 
 def _process_inputs(args, kwargs) -> Any:
-  """A function that normalizes the representation of the ``args`` and ``kwargs``
-  for the ``inputs`` column."""
+  """A function that normalizes the representation of the ``args`` and
+  ``kwargs`` for the ``inputs`` column.
+  """
   if args and kwargs:
     input_values = (*args, kwargs)
   elif args and not kwargs:
@@ -399,13 +554,13 @@ def _render_table(
     console_extras: Optional[Mapping[str, Any]],
     table_kwargs: Mapping[str, Any],
     column_kwargs: Mapping[str, Any],
+    non_params_cols: List[str],
 ) -> str:
   """A function that renders a Table to a string representation using rich."""
   console_kwargs = {'force_terminal': True, 'force_jupyter': False}
   if console_extras is not None:
     console_kwargs.update(console_extras)
 
-  non_params_cols = 4
   rich_table = rich.table.Table(
       show_header=True,
       show_lines=True,
@@ -414,10 +569,8 @@ def _render_table(
       **table_kwargs,
   )
 
-  rich_table.add_column('path', **column_kwargs)
-  rich_table.add_column('module', **column_kwargs)
-  rich_table.add_column('inputs', **column_kwargs)
-  rich_table.add_column('outputs', **column_kwargs)
+  for c in non_params_cols:
+    rich_table.add_column(c, **column_kwargs)
 
   for col in table.collections:
     rich_table.add_column(col, **column_kwargs)
@@ -450,17 +603,20 @@ def _render_table(
     rich_table.add_row(
         path_repr,
         row.module_type.__name__ + method_repr,
-        _as_yaml_str(
-            _summary_tree_map(_maybe_render, _normalize_structure(row.inputs))
-        ),
-        _as_yaml_str(
-            _summary_tree_map(_maybe_render, _normalize_structure(row.outputs))
+        *(
+            _as_yaml_str(
+                _summary_tree_map(
+                    _maybe_render, _normalize_structure(getattr(row, c))
+                )
+            )
+            for c in non_params_cols[2:]
         ),
         *collections_size_repr,
     )
 
   # add footer with totals
-  rich_table.columns[non_params_cols - 1].footer = rich.text.Text.from_markup(
+  n_non_params_cols = len(non_params_cols)
+  rich_table.columns[n_non_params_cols - 1].footer = rich.text.Text.from_markup(
       'Total', justify='right'
   )
 
@@ -475,7 +631,7 @@ def _render_table(
 
   # add totals to footer
   for i, col in enumerate(table.collections):
-    rich_table.columns[non_params_cols + i].footer = _size_and_bytes_repr(
+    rich_table.columns[n_non_params_cols + i].footer = _size_and_bytes_repr(
         *collection_total[col]
     )
 
@@ -579,6 +735,21 @@ def _get_value_representation(x: Any) -> _ValueRepresentation:
     return _ArrayRepresentation.from_array(x)
   except:
     return _ObjectRepresentation(x)
+
+
+def _from_value_representation(x: _ValueRepresentation) -> Any:
+  if isinstance(x, _ArrayRepresentation):
+    return jax.ShapeDtypeStruct(x.shape, x.dtype)
+
+  elif isinstance(x, _PartitionedArrayRepresentation):
+    return jax.ShapeDtypeStruct(
+        x.array_representation.shape, x.array_representation.dtype
+    )
+
+  elif isinstance(x, _ObjectRepresentation):
+    return x.obj
+
+  raise TypeError(x, type(x))
 
 
 def _represent_tree(x):